small cleanup, rename to "no-force"
