+++ lib/rules/no-force.js
